refactor(capabilities): extract shared TestChassis test fixture (issue 785)
Every cap in aether-capabilities declared the same 8-line
`impl Chassis for TestChassis` block in its test module — 11 byte-identical
sites. Extract the canonical declaration into a new
`#[cfg(test)] pub(crate) mod test_chassis;` so each test module just
imports `crate::test_chassis::TestChassis;` instead.

Also drops the now-unused `Chassis`, `BuiltChassis`, `NeverDriver`, and
(where applicable) `BootError` imports from the affected test modules.
No behavior change — same `Chassis` impl, same `Builder` boot path.
