+++ crates/aether-capabilities/src/log.rs
@@ -118,23 +118,13 @@ mod native {
         use std::time::Duration;
 
         use super::{BootError, LogBatch, LogCapability};
+        use crate::test_chassis::TestChassis;
         use aether_actor::Actor;
         use aether_data::Kind;
         use aether_kinds::LogEvent;
-        use aether_substrate::chassis::Chassis;
-        use aether_substrate::chassis::builder::{Builder, BuiltChassis, NeverDriver};
+        use aether_substrate::chassis::builder::Builder;
         use aether_substrate::mail::mailer::Mailer;
         use aether_substrate::mail::registry::{MailboxEntry, Registry};
-
-        struct TestChassis;
-        impl Chassis for TestChassis {
-            const PROFILE: &'static str = "test";
-            type Driver = NeverDriver;
-            type Env = ();
-            fn build(_env: Self::Env) -> Result<BuiltChassis<Self>, BootError> {
-                unreachable!("TestChassis is driven by Builder::new directly in unit tests")
-            }
-        }
 
         fn fresh_substrate() -> (Arc<Registry>, Arc<Mailer>) {
             {
