+++ crates/aether-capabilities/src/rpc/server.rs
@@ -681,10 +681,9 @@ pub fn rpc_server_mailbox_id() -> aether_data::MailboxId {
 mod tests {
     use super::*;
     use crate::rpc::wire::{Hello, HelloAck, PeerKind, WIRE_VERSION, WireFrame};
+    use crate::test_chassis::TestChassis;
     use aether_codec::frame::{read_frame, write_frame};
-    use aether_substrate::chassis::Chassis;
-    use aether_substrate::chassis::builder::{Builder, BuiltChassis, NeverDriver};
-    use aether_substrate::chassis::error::BootError;
+    use aether_substrate::chassis::builder::Builder;
     use aether_substrate::handle_store::HandleStore;
     use aether_substrate::mail::mailer::Mailer;
     use aether_substrate::mail::outbound::HubOutbound;
@@ -692,16 +691,6 @@ mod tests {
     use std::net::TcpStream;
     use std::sync::Arc;
     use std::time::Duration;
-
-    struct TestChassis;
-    impl Chassis for TestChassis {
-        const PROFILE: &'static str = "test";
-        type Driver = NeverDriver;
-        type Env = ();
-        fn build(_env: Self::Env) -> Result<BuiltChassis<Self>, BootError> {
-            unreachable!("TestChassis is driven by Builder::new directly in unit tests")
-        }
-    }
 
     fn fresh_substrate() -> (Arc<Registry>, Arc<Mailer>) {
         let registry = Arc::new(Registry::new());
